fix ending conference if it was not created

+++ app/BUILD/src/DeskPRO/Bundle/VoiceBundle/Plivo/PlivoAdapter.php
@@ -363,7 +363,12 @@ class PlivoAdapter implements VoiceProviderInterface
         }
 
         $conference = $this->getConference($account, $phoneCall->getConferenceName());
-        $conference->delete();
+        if ($conference) {
+            try {
+                $conference->delete();
+            } catch (\Exception $e) {
+            }
+        }
 
         foreach ($phoneCall->getParticipants() as $participant) {
             try {
